tests : refactor some classes
